LayoutTests/imported/w3c/web-platform-tests/dom/events/EventListener-addEventListener.sub.window-expected.txt


PASS EventListener.addEventListener doesn't throw when a cross origin object is passed in.
